+++ .github/workflows/deploy.yml
@@ -1,10 +1,12 @@
 name: KIYO CI
 
-
 on:
   push:
     branches:
       - 'master'
+      - 'develop'
+  pull_request:
+    branches:
       - 'develop'
 env:
   S3_BUCKET_NAME: kiyobackend
