+++ src/main/java/project/kiyobackend/QnA/domain/QnA.java
@@ -2,7 +2,7 @@ package project.kiyobackend.QnA.domain;
 
 import lombok.Getter;
 import project.kiyobackend.user.domain.User;
-import project.kiyobackend.util.jpa.JpaBaseEntity;
+import project.kiyobackend.common.util.jpa.JpaBaseEntity;
 
 import javax.persistence.*;
