--- src/main/java/project/kiyobackend/auth/entity/Token.java
@@ -1,4 +1,0 @@
-package project.kiyobackend.auth.entity;
-
-public class Token {
-}
