+++ src/main/java/project/kiyobackend/auth/repository/OAuth2AuthorizationRequestBasedOnCookieRepository.java
@@ -3,7 +3,7 @@ package project.kiyobackend.auth.repository;
 import com.nimbusds.oauth2.sdk.util.StringUtils;
 import org.springframework.security.oauth2.client.web.AuthorizationRequestRepository;
 import org.springframework.security.oauth2.core.endpoint.OAuth2AuthorizationRequest;
-import project.kiyobackend.util.auth.CookieUtil;
+import project.kiyobackend.common.util.auth.CookieUtil;
 
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
