+++ src/main/java/project/kiyobackend/common/config/AwsS3Config.java
@@ -1,4 +1,4 @@
-package project.kiyobackend.util;
+package project.kiyobackend.common.config;
 
 import com.amazonaws.auth.AWSStaticCredentialsProvider;
 import com.amazonaws.auth.BasicAWSCredentials;
