+++ src/main/java/project/kiyobackend/common/util/auth/CookieUtil.java
@@ -1,4 +1,4 @@
-package project.kiyobackend.util.auth;
+package project.kiyobackend.common.util.auth;
 
 import org.springframework.util.SerializationUtils;
