+++ src/main/java/project/kiyobackend/common/util/jpa/JpaBaseEntity.java
@@ -1,4 +1,4 @@
-package project.kiyobackend.util.jpa;
+package project.kiyobackend.common.util.jpa;
 
 import lombok.Getter;
 import org.springframework.data.annotation.CreatedDate;
@@ -8,7 +8,6 @@ import org.springframework.data.jpa.domain.support.AuditingEntityListener;
 import javax.persistence.Column;
 import javax.persistence.EntityListeners;
 import javax.persistence.MappedSuperclass;
-import javax.persistence.PrePersist;
 import java.time.LocalDateTime;
 
 @MappedSuperclass
