+++ src/main/java/project/kiyobackend/review/domain/domain/Review.java
@@ -3,7 +3,7 @@ package project.kiyobackend.review.domain.domain;
 import lombok.Getter;
 import project.kiyobackend.store.domain.domain.store.Store;
 import project.kiyobackend.user.domain.User;
-import project.kiyobackend.util.jpa.JpaBaseEntity;
+import project.kiyobackend.common.util.jpa.JpaBaseEntity;
 
 import javax.persistence.*;
 import java.util.ArrayList;
