+++ src/main/java/project/kiyobackend/store/domain/domain/menu/MenuOption.java
@@ -3,7 +3,7 @@ package project.kiyobackend.store.domain.domain.menu;
 import lombok.AccessLevel;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
-import project.kiyobackend.util.jpa.JpaBaseEntity;
+import project.kiyobackend.common.util.jpa.JpaBaseEntity;
 
 import javax.persistence.*;
