+++ src/main/java/project/kiyobackend/store/query/dto/StoreRequestDto.java
@@ -1,4 +1,4 @@
-package project.kiyobackend.store.query;
+package project.kiyobackend.store.query.dto;
 
 import lombok.Data;
